RESOLVED - issue OPENFLEXO-755: ajout d'un deletion scheme https://bugs.openflexo.com/browse/OPENFLEXO-755 Controller was not properly deleted 	•  Controller did not remove its observer upon deletion 	•  Some NPE were not prevented

+++ flexodesktop/GUI/flexointerfacebuilder/src/main/java/org/openflexo/fib/controller/FIBController.java
@@ -153,7 +153,9 @@ public class FIBController extends Observable implements BindingEvaluationContex
 
 	public void delete() {
 		if (!deleted) {
-			getRootView().delete();
+			if (getRootView() != null) {
+				getRootView().delete();
+			}
 			// Next for-block should not be necessary because deletion is recursive, but just to be sure
 			for (FIBView<?, ?> view : new ArrayList<FIBView<?, ?>>(views.values())) {
 				view.delete();
